+++ blawx/views.py
@@ -219,6 +219,20 @@ class ScenarioEditor(PermissionRequiredMixin, generic.DetailView):
         context['ruledoc_list'] = RuleDoc.objects.all()
         return context
 
+class ScenarioEditorGCWeb(PermissionRequiredMixin, generic.DetailView):
+    permission_required = "blawx.view_blawxtest"
+    template_name = "blawx/scenario_editor_gcweb.html"
+    model = BlawxTest
+
+    def get_object(self):
+        return BlawxTest.objects.get(ruledoc=RuleDoc.objects.get(pk=self.kwargs['ruledoc']),test_name=self.kwargs['test_name'])
+    
+    def get_context_data(self, **kwargs):
+        context = super(ScenarioEditorGCWeb, self).get_context_data(**kwargs)
+        context['ruledoc_list'] = RuleDoc.objects.all()
+        return context
+
+
 class TestCreateView(PermissionRequiredMixin, CreateView):
     permission_required = "blawx.add_blawxtest_to_ruledoc"
     model = BlawxTest